Uploaded the files I worked on
It operates like a Turing machine. It can move left and right, it has conditionals, it has goto* statements (I call them flags in the code). I am thinking of calling it "Capture the Flag" in honor of the flags acting like GOTO statements. Not sure if it is truly Turing complete but it seems like it is. In the program I have a "Hello World" program,  a "Recursive remove all 1s function" to demonstrate flags, conditionals, reading and writing and an infinite loop. It uses a linked list as the turning machine tape so I do not have to worry about resizing